fix smallInt for plan usage

+++ Modules/Subscriptions/Database/Migrations/2023_04_03_115725_create_plan_subscription_usage.php
@@ -21,7 +21,7 @@ return new class extends Migration
                 ->constrained('plan_features')
                 ->cascadeOnDelete()
                 ->cascadeOnUpdate();
-            $table->unsignedSmallInteger('used');
+            $table->unsignedBigInteger('used');
             $table->dateTime('valid_until')->nullable();
             $table->timestamps();
         });
